fix(ci): python-full-coverage — split RepoRef out of github_client.py to clear 1500-line maintainability limit
github_client.py had grown to 1518 lines, tripping
test_first_party_code_files_stay_under_line_limit (max 1500). Extract the
self-contained RepoRef value object plus its forge-host maps and
_strip_bare_slug_git_suffix helper into a new github_client_ref module,
matching the existing github_client_* decomposition pattern, and re-export
RepoRef for compatibility. Behavior-preserving move; existing tests cover
the new module at 100%.
